apps/api/pyproject.toml
[project]
name = "api"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dev = "uvicorn src.api.main:app --reload --port 8000"